+++ JGestion/test/jgestion/test/JXTreeTableDemo.java
@@ -263,7 +263,7 @@ public class JXTreeTableDemo extends javax.swing.JDialog {
                     }
                     List<DetalleCajaMovimientos> mviList = getMovimientosVariosIngresos(month);
                     for (DetalleCajaMovimientos o : mviList) {
-                        BigDecimal importe = BigDecimal.valueOf(o.getMonto());
+                        BigDecimal importe = o.getMonto();
                         if (o.getUnidadDeNegocio() == null) {
                             ingresosSinUNDniCuenta = ingresosSinUNDniCuenta.add(importe);
                             continue;
@@ -321,7 +321,7 @@ public class JXTreeTableDemo extends javax.swing.JDialog {
                     List<DetalleCajaMovimientos> mveList = getMovimientosVariosEgresos(month);
                     for (DetalleCajaMovimientos o : mveList) {
                         // LOS MV EGRESOS YA son Negativos!!!!!
-                        BigDecimal importe = BigDecimal.valueOf(o.getMonto());
+                        BigDecimal importe = o.getMonto();
                         if (o.getUnidadDeNegocio() == null) {
                             egresosSinUNDniCuenta = egresosSinUNDniCuenta.add(importe);
                             continue;
